// apps/admin-frontend/lib/permissions/module.enum.ts
export enum Module {
  STUDENT_FEES = "STUDENT_FEES",
  FEE_REPORTS = "FEE_REPORTS",
  PAYMENT_HISTORY = "PAYMENT_HISTORY",
  STUDENT_ADMISSIONS = "STUDENT_ADMISSIONS",
  STUDENT_REGISTRATION = "STUDENT_REGISTRATION",
  PARENT_DETAILS = "PARENT_DETAILS",
  TEACHER_MANAGEMENT = "TEACHER_MANAGEMENT",
  EXAM_SETTINGS = "EXAM_SETTINGS",
  ACADEMIC_YEAR = "ACADEMIC_YEAR",
  USER_MANAGEMENT = "USER_MANAGEMENT",
  ATTENDANCE = "ATTENDANCE",
  PAYROLL = "PAYROLL",
}